backend/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-socketio[client]==5.9.0
faker==19.13.0
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2